Add shared state, move 105 to own visitor file, move utility functionality out to new file visitor_utility.py
