+++ flake8_trio/visitors/flake8triovisitor.py
@@ -4,34 +4,57 @@ from __future__ import annotations
 
 import ast
 import re
-from typing import TYPE_CHECKING, Any, Union, cast
+from abc import ABC
+from typing import TYPE_CHECKING, Any, Union
 
 from ..base import Error, Statement
 
 if TYPE_CHECKING:
-    from argparse import Namespace
     from collections.abc import Iterable
+
+    from ..runner import SharedState
 
     HasLineCol = Union[
         ast.expr, ast.stmt, ast.arg, ast.excepthandler, ast.alias, Statement
     ]
 
 
-class Flake8TrioVisitor(ast.NodeVisitor):
+class Flake8TrioVisitor(ast.NodeVisitor, ABC):
     # abstract attribute by not providing a value
     error_codes: dict[str, str]
 
-    def __init__(self, options: Namespace, _problems: list[Error]):
+    def __init__(self, shared_state: SharedState):
         super().__init__()
         assert self.error_codes, "subclass must define error_codes"
-        self._problems = _problems
-        self.suppress_errors = False
-        self.options = options
         self.outer: dict[ast.AST, dict[str, Any]] = {}
         self.novisit = False
-        self._library: tuple[str, ...] = ()
-        if self.options.anyio:
-            self._library = ("anyio",)
+        self.__state = shared_state
+
+        self.options = self.__state.options
+        self.typed_calls = self.__state.typed_calls
+
+        # mark variables that shouldn't be saved/loaded in self.get_state
+        self.nocopy = {
+            "_Flake8TrioVisitor__state",
+            "error_codes",
+            "nocopy",
+            "novisit",
+            "options",
+            "outer",
+            "typed_calls",
+        }
+
+        self.suppress_errors = False
+
+    # `variables` can be saved/loaded, but need a setter to not clear the reference
+    @property
+    def variables(self):
+        return self.__state.variables
+
+    @variables.setter
+    def variables(self, value):
+        self.__state.variables.clear()
+        self.__state.variables.update(value)
 
     def visit(self, node: ast.AST):
         """Visit a node."""
@@ -80,7 +103,7 @@ class Flake8TrioVisitor(ast.NodeVisitor):
             return
 
         if not self.suppress_errors:
-            self._problems.append(
+            self.__state.problems.append(
                 Error(
                     # 7 == len('TRIO...'), so alt messages raise the original code
                     error_code[:7],
@@ -93,12 +116,8 @@ class Flake8TrioVisitor(ast.NodeVisitor):
 
     def get_state(self, *attrs: str, copy: bool = False) -> dict[str, Any]:
         if not attrs:
-            attrs = cast(
-                tuple[str, ...],
-                set(self.__dict__.keys())
-                - {"_problems", "options", "outer", "novisit", "error_codes"},
-                # TODO: write something clean so don't need to hardcode
-            )
+            # get all attributes, unless they're marked as nocopy
+            attrs = tuple(set(self.__dict__.keys()) - self.nocopy)
         res: dict[str, Any] = {}
         for attr in attrs:
             value = getattr(self, attr)
@@ -116,7 +135,9 @@ class Flake8TrioVisitor(ast.NodeVisitor):
     def save_state(self, node: ast.AST, *attrs: str, copy: bool = False):
         state = self.get_state(*attrs, copy=copy)
         if node in self.outer:
-            self.outer[node].update(state)
+            # not currently used, and not gonna bother adding dedicated test
+            # visitors atm
+            self.outer[node].update(state)  # pragma: no cover
         else:
             self.outer[node] = state
 
@@ -124,22 +145,16 @@ class Flake8TrioVisitor(ast.NodeVisitor):
         for b in body:
             yield from ast.walk(b)
 
-    def visit_Import(self, node: ast.Import):
-        for alias in node.names:
-            name = alias.name
-            if (
-                name in ("trio", "anyio")
-                and name not in self._library
-                and alias.asname is None
-            ):
-                self._library = self._library + (name,)
-
     @property
     def library(self) -> tuple[str, ...]:
-        return self._library if self._library else ("trio",)
+        return self.__state.library if self.__state.library else ("trio",)
 
     @property
     def library_str(self) -> str:
         if len(self.library) == 1:
             return self.library[0]
         return "[" + "|".join(self.library) + "]"
+
+    def add_library(self, name) -> None:
+        if name not in self.__state.library:
+            self.__state.library = self.__state.library + (name,)
